isolated method to fetch all RFCs
